refactor wind speed direction component name
