+++ src/components/widgets/speed-direction-widget.tsx
